requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
alembic
pydantic
pydantic[email]
google-generativeai
playwright
